+++ src/revise/prompts/commit_body.rs
@@ -11,13 +11,11 @@ pub fn inquire_commit_body() -> ReviseResult<Option<String>> {
         .with_formatter(&|submission| {
             let char_count = submission.chars().count();
             if char_count == 0 {
-                String::from("<skipped>")
+                "<skipped>".to_string()
             } else if char_count <= 20 {
                 submission.into()
             } else {
-                let mut substr: String = submission.chars().take(17).collect();
-                substr.push_str("...");
-                substr
+                format!("{}...", &submission[..17])
             }
         })
         .with_render_config(
@@ -26,9 +24,9 @@ pub fn inquire_commit_body() -> ReviseResult<Option<String>> {
             ),
         )
         .prompt()?;
-    if ans.is_empty() {
-        Ok(None)
-    } else {
-        Ok(Some(ans))
+
+    match &*ans {
+        "<skipped>" | "" => Ok(None),
+        _ => Ok(Some(ans)),
     }
 }