+++ src/revise/prompts/commit_breaking.rs
@@ -1,4 +1,7 @@
-use inquire::Editor;
+use inquire::{
+    ui::{Color, RenderConfig, Styled},
+    Editor,
+};
 
 use crate::error::ReviseResult;
 
@@ -8,19 +11,21 @@ pub fn inquire_commit_breaking() -> ReviseResult<Option<String>> {
         .with_formatter(&|submission| {
             let char_count = submission.chars().count();
             if char_count == 0 {
-                String::from("<skipped>")
+                "<skipped>".to_string()
             } else if char_count <= 20 {
                 submission.into()
             } else {
-                let mut substr: String = submission.chars().take(17).collect();
-                substr.push_str("...");
-                substr
+                format!("{}...", &submission[..17])
             }
         })
+        .with_render_config(
+            RenderConfig::default().with_canceled_prompt_indicator(
+                Styled::new("<skipped>").with_fg(Color::DarkYellow),
+            ),
+        )
         .prompt()?;
-    if ans.is_empty() {
-        Ok(None)
-    } else {
-        Ok(Some(ans))
+    match &*ans {
+        "<skipped>" | "" => Ok(None),
+        _ => Ok(Some(ans)),
     }
 }