+++ src/revise/prompts/commit_edit.rs
@@ -6,7 +6,7 @@ use inquire::{
 use crate::error::ReviseResult;
 
 pub fn inquire_commit_edit(messages: &str) -> ReviseResult<String> {
-    let msg = "You Really want to edit this commit by yourself?";
+    let msg = "You Really want to edit this commit manually?";
     let ans = Editor::new("")
         .with_predefined_text(messages)
         .with_render_config(